Experimental: add Epoll.h, Libuv.h, pass most tests
