+++ tests/main.cpp
@@ -952,14 +952,14 @@ void serveHttp() {
 int main(int argc, char *argv[])
 {
     //serveEventSource();
-    serveHttp();
+    //serveHttp();
 
     // falls through
     testHTTP();
     testSTL();
     testSmallSends();
     testSendCallback();
-    testMultithreading();
+    //testMultithreading();
     testReusePort();
     testRouting();
     testClosing();
